Replacing old Java date handling with newer java.time code
The SimpleDateFormatter I was using in the last commit wasn't
sufficiently nimble for the different types of timestamps we're seeing
in "real" Stories.  Instead of using jodatime with its IsoDateFormatter,
I think the latest Java 8 java.time package code should be able to
support our DateFormat requirements with the Instant class.