Fix tests that should run only when fixture exists
The `fixtures` array is only initialized within the `it` call, so refactor
by adding helper function `validateIfFixtureExists`.

Before the fix, not all tests ran to success. To address this remove the
v0.5.5 test files that we can't fix without modifying the 0.5.5 schemas,
and modify the v0.5.6 schemas that we can.